refactor(RacingCar): 자동차 객체 이름 속성 추가
